+++ src/Mc.PowerShell/Mc.psm1
@@ -141,6 +141,7 @@ Export-ModuleMember -Function @(
     'Show-McSortMenu'
     'Get-McPanelCurrent'
     'Get-McPanelLeafName'
+    'Test-McXmlFile'
     'Move-McPanelCursor'
     'Set-McPanelCursor'
     'Set-McPanelLocation'